+++ CHANGELOG.md
@@ -6,9 +6,13 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 ### Added
-- Project: The zebra bra
-- Project: The sage pants
 ### Changed
+
+## [0.4.0] - 2019-08-31
+### Added
+- Production: The sage pants (to ep38)
+- Production: Yusha now (to ep42)
+- Production: Zebra bra
 
 ## [0.3.2] - 2019-08-01
 ### Added
@@ -42,7 +46,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Setup file for python.
 - Submodule: storybuilder is a library to build a story.
 
-[Unreleased]: https://github.com/nagisc007/prj_estar/compare/v0.3.2...HEAD
+[Unreleased]: https://github.com/nagisc007/prj_estar/compare/v0.4.0...HEAD
+[0.4.0]: https://github.com/nagisc007/prj_estar/releases/v0.4.0
 [0.3.2]: https://github.com/nagisc007/prj_estar/releases/v0.3.2
 [0.3.1]: https://github.com/nagisc007/prj_estar/releases/v0.3.1
 [0.3.0]: https://github.com/nagisc007/prj_estar/releases/v0.3.0